+++ src/helpers/helpers.ts
@@ -2,10 +2,7 @@ import * as Immutable from 'immutable';
 
 export class Helpers {
     static overrideContructor(obj: any) {
-        if (obj !== null
-            && typeof (obj) === 'object'
-            && !Immutable.Map.isMap(obj)
-            && !Immutable.Iterable.isIterable(obj)) {
+        if (Helpers.isObject(obj) && !Helpers.isImmutable(obj)) { // from ImmutableJs 4 breaking change isIterable => isCollection
             if (obj.constructor === Array) {
                 for (let i = 0; i < obj.length; i++) {
                     this.overrideContructor(obj[i]);
@@ -18,4 +15,15 @@ export class Helpers {
             }
         }
     };
+
+    static isImmutable(obj: any) {
+        return Immutable.Map.isMap(obj) || Immutable.Iterable.isIterable(obj);
+    }
+
+    static isObject(obj: any) {
+        return obj !== null
+            && typeof (obj) === 'object'
+            && !Immutable.Map.isMap(obj)
+            && !Immutable.Iterable.isIterable(obj);
+    }
 }