+++ src/index.ts
@@ -7,10 +7,20 @@ import { StateHistory } from "./state/history";
 import { Store } from "./store/store";
 
 export class ReactState {
-    static init(domRender: (history: History) => void, initialState: any, collectHistory: boolean = true, storeHistoryItems: number = 100) {
+    static init(
+        domRender: (history: History) => void,
+        initialState: any,
+        isProd: boolean = false,
+        collectHistory: boolean = true,
+        storeHistoryItems: number = 100
+    ) {
         const store = new Store(new State(initialState));
-        new StateHistory(store, collectHistory, storeHistoryItems).init();
-        (<any>window).state = StateHistory;
+        new StateHistory(store, collectHistory, storeHistoryItems).init(initialState);
+
+        if(!isProd) {
+            (<any>window).state = StateHistory;
+        }
+
         const routerHistory = new RouterState(store).init();
 
         Store.store = store;