+++ src/store/store.ts
@@ -1,3 +1,4 @@
+import { Clear, ClearSignature } from './clear';
 import * as Rx from 'rxjs';
 
 import { Initialize, InitializeSignature } from './initialize';
@@ -5,7 +6,7 @@ import { Select, SelectSignature } from './select';
 import { Update, UpdateSignature } from './update';
 
 export class Store<T> extends Rx.Observable<T> implements Rx.Observer<any> {
-    static store: any;
+    static store: Store<any>;
     constructor(state: Rx.Observable<any>) {
         super();
 
@@ -20,6 +21,7 @@ export class Store<T> extends Rx.Observable<T> implements Rx.Observer<any> {
 
     update: UpdateSignature<T> = Update.bind(this);
     initialize: InitializeSignature<T> = Initialize.bind(this);
+    clear: ClearSignature = Clear.bind(this);
 
     lift<R>(operator: Rx.Operator<T, R>): Store<R> {
         const store = new Store<R>(this);